updating sitemap to allow searching within a particular portion of the tree
git-svn-id: http://svn.concrete5.org/svn/concrete5@790 b0551a0c-1e16-4222-a7d5-975db1aca215

+++ trunk/web/concrete/js/ccm.sitemap.js
@@ -33,7 +33,6 @@ showPageMenu = function(obj, e) {
 		html += '<div class="ccm-menu-l"><div class="ccm-menu-r">';
 		html += "<ul>";
 		
-		
 		if (obj.cAlias == 'LINK' || obj.cAlias == 'POINTER') {
 		
 			html += '<li><a class="ccm-icon" id="menuVisit' + obj.cID + '" href="' + CCM_REL + '/index.php?cID=' + obj.cID + '"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/window_new.png)">' + ccmi18n_sitemap.visitExternalLink + '<\/span><\/a><\/li>';
@@ -50,20 +49,27 @@ showPageMenu = function(obj, e) {
 		
 		} else {
 		
-
-			html += '<li><a class="ccm-icon" id="menuVisit' + obj.cID + '" href="' + CCM_REL + '/index.php?cID=' + obj.cID + '"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/window_new.png)">' + ccmi18n_sitemap.visitPage + '<\/span><\/a><\/li>';
-			html += '<li><a class="ccm-icon" dialog-width="640" dialog-height="310" dialog-modal="false" dialog-title="' + ccmi18n_sitemap.pageProperties + '" id="menuProperties' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=edit_metadata"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/edit_small.png)">' + ccmi18n_sitemap.pageProperties + '<\/span><\/a><\/li>';
-			html += '<li><a class="ccm-icon" dialog-width="640" dialog-height="310" dialog-modal="false" dialog-title="' + ccmi18n_sitemap.setPagePermissions + '" id="menuPermissions' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=edit_permissions"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/permissions_small.png)">' + ccmi18n_sitemap.setPagePermissions + '<\/span><\/a><\/li>';
-			html += '<li><a class="ccm-icon" dialog-width="680" dialog-height="420" dialog-modal="false" dialog-title="' + ccmi18n_sitemap.pageDesign + '" id="menuDesign' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=set_theme"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/design_small.png)">' + ccmi18n_sitemap.pageDesign + '<\/span><\/a><\/li>';
-			html += '<li><a class="ccm-icon" dialog-width="640" dialog-height="340" dialog-modal="false" dialog-title="' + ccmi18n_sitemap.pageVersions + '" id="menuVersions' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/versions.php?rel=SITEMAP&cID=' + obj.cID + '"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/versions_small.png)">' + ccmi18n_sitemap.pageVersions + '<\/span><\/a><\/li>';
-			html += '<li><a class="ccm-icon" id="menuDelete' + obj.cID + '" href="javascript:deletePage(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/delete_small.png)">' + ccmi18n_sitemap.deletePage + '<\/span><\/a><\/li>';
-			html += '<li class=\"header\"><\/li>';
-			html += '<li><a class="ccm-icon" id="menuReorder' + obj.cID + '" href="javascript:activateReorder(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/up_down.png)">' + ccmi18n_sitemap.reorderPage + '<\/span><\/a><\/li>';
-			html += '<li><a class="ccm-icon" id="menuMoveCopy' + obj.cID + '" href="javascript:activateMoveCopy(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/up_down.png)">' + ccmi18n_sitemap.moveCopyPage + '<\/span><\/a><\/li>';
-			html += '<li class=\"header\"><\/li>';
-			html += '<li><a class="ccm-icon ccm-icon-sitemap-search" id="menuSearch' + obj.cID + '" href="javascript:searchSubPages(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/magnifying.png)">' + ccmi18n_sitemap.searchPages + '<\/span><\/a><\/li>';
-			html += '<li><a class="ccm-icon" dialog-width="680" dialog-modal="false" dialog-height="440" dialog-title="' + ccmi18n_sitemap.addPage + '" id="menuSubPage' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=add"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/add.png)">' + ccmi18n_sitemap.addPage + '<\/span><\/a><\/li>';
-			html += '<li><a class="ccm-icon" dialog-width="350" dialog-modal="false" dialog-height="160" dialog-title="' + ccmi18n_sitemap.addExternalLink + '" dialog-modal="false" id="menuLink' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=add_external"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/add.png)">' + ccmi18n_sitemap.addExternalLink + '<\/span><\/a><\/li>';
+			if (CCM_SITEMAP_MODE == 'full') {
+				html += '<li><a class="ccm-icon" id="menuVisit' + obj.cID + '" href="' + CCM_REL + '/index.php?cID=' + obj.cID + '"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/window_new.png)">' + ccmi18n_sitemap.visitPage + '<\/span><\/a><\/li>';
+				html += '<li><a class="ccm-icon" dialog-width="640" dialog-height="310" dialog-modal="false" dialog-title="' + ccmi18n_sitemap.pageProperties + '" id="menuProperties' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=edit_metadata"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/edit_small.png)">' + ccmi18n_sitemap.pageProperties + '<\/span><\/a><\/li>';
+				html += '<li><a class="ccm-icon" dialog-width="640" dialog-height="310" dialog-modal="false" dialog-title="' + ccmi18n_sitemap.setPagePermissions + '" id="menuPermissions' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=edit_permissions"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/permissions_small.png)">' + ccmi18n_sitemap.setPagePermissions + '<\/span><\/a><\/li>';
+				html += '<li><a class="ccm-icon" dialog-width="680" dialog-height="420" dialog-modal="false" dialog-title="' + ccmi18n_sitemap.pageDesign + '" id="menuDesign' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=set_theme"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/design_small.png)">' + ccmi18n_sitemap.pageDesign + '<\/span><\/a><\/li>';
+				html += '<li><a class="ccm-icon" dialog-width="640" dialog-height="340" dialog-modal="false" dialog-title="' + ccmi18n_sitemap.pageVersions + '" id="menuVersions' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/versions.php?rel=SITEMAP&cID=' + obj.cID + '"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/versions_small.png)">' + ccmi18n_sitemap.pageVersions + '<\/span><\/a><\/li>';
+				html += '<li><a class="ccm-icon" id="menuDelete' + obj.cID + '" href="javascript:deletePage(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/delete_small.png)">' + ccmi18n_sitemap.deletePage + '<\/span><\/a><\/li>';
+				html += '<li class=\"header\"><\/li>';
+				html += '<li><a class="ccm-icon" id="menuReorder' + obj.cID + '" href="javascript:activateReorder(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/up_down.png)">' + ccmi18n_sitemap.reorderPage + '<\/span><\/a><\/li>';
+				html += '<li><a class="ccm-icon" id="menuMoveCopy' + obj.cID + '" href="javascript:activateMoveCopy(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/up_down.png)">' + ccmi18n_sitemap.moveCopyPage + '<\/span><\/a><\/li>';
+				html += '<li class=\"header\"><\/li>';
+				if (obj.cNumChildren > 0 && obj.cID > 1) {
+					html += '<li><a class="ccm-icon ccm-icon-sitemap-search" id="menuSearch' + obj.cID + '" href="javascript:searchSubPages(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/magnifying.png)">' + ccmi18n_sitemap.searchPages + '<\/span><\/a><\/li>';
+				}
+				html += '<li><a class="ccm-icon" dialog-width="680" dialog-modal="false" dialog-height="440" dialog-title="' + ccmi18n_sitemap.addPage + '" id="menuSubPage' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=add"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/add.png)">' + ccmi18n_sitemap.addPage + '<\/span><\/a><\/li>';
+				html += '<li><a class="ccm-icon" dialog-width="350" dialog-modal="false" dialog-height="160" dialog-title="' + ccmi18n_sitemap.addExternalLink + '" dialog-modal="false" id="menuLink' + obj.cID + '" href="' + CCM_TOOLS_PATH + '/edit_collection_popup.php?rel=SITEMAP&cID=' + obj.cID + '&ctask=add_external"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/add.png)">' + ccmi18n_sitemap.addExternalLink + '<\/span><\/a><\/li>';
+			} else {
+				// we only get here if there are enough pages under here to search
+				html += '<li><a class="ccm-icon ccm-icon-sitemap-search" id="menuSearch' + obj.cID + '" href="javascript:searchSubPages(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/magnifying.png)">' + ccmi18n_sitemap.searchPages + '<\/span><\/a><\/li>';
+				html += '<li><a class="ccm-icon" id="menuMoveCopy' + obj.cID + '" href="javascript:selectMoveCopyTarget(' + obj.cID + ')"><span style="background-image: url(' + CCM_IMAGE_PATH + '/icons/up_down.png)">' + ccmi18n_sitemap.moveCopyPage + '<\/span><\/a><\/li>';
+			}
 
 		} 
 		
@@ -133,6 +139,17 @@ deactivateMoveCopy = function() {
 	tr_moveCopyMode = false;
 	CCM_SITEMAP_MODE = 'full';
 	hideSitemapMessage();
+}
+
+searchSubPages = function(cID) {
+	if (ccm_animEffects) {
+		$("#ccm-tree-search" + cID).fadeIn(200, function() {
+			$("#ccm-tree-search" + cID + " input").get(0).focus();
+		});
+	} else {
+		$("#ccm-tree-search" + cID).show();
+		$("#ccm-tree-search" + cID + " input").get(0).focus();
+	}
 }
 
 activateReorder = function(cID) {
@@ -259,9 +276,9 @@ parseTree = function(node, nodeID, deactivateSubNodes) {
 		html += '<li tree-node-type="' + treeNodeType + '" class="tree-node ' + typeClass + ' tree-branch' + nodeID + '" id="tree-node' + node[i].id + '"' + customIconSrc + '>';
 		if (node[i].numSubpages > 0 && (!deactivateSubNodes)) {
 			var subPageStr = (node[i].id == 1) ? '' : ' (' + node[i].numSubpages + ')';
-			html += '<img src="' + CCM_IMAGE_PATH + '/spacer.gif" width="16" height="16" class="handle ' + moveableClass + '" /><a href="javascript:toggleSub(' + node[i].id + ')"><img src="' + CCM_IMAGE_PATH + '/dashboard/plus.jpg" width="9" height="9" class="tree-plus" id="tree-collapse' + node[i].id + '" /><\/a><div rel="' + CCM_REL + '/index.php?cID=' + node[i].id + '" class="' + labelClass + '" tree-node-alias="' + cAlias + '" tree-node-canwrite="' + canWrite + '" tree-node-title="' + escape(node[i].cvName) + '" id="tree-label' + node[i].id + '"><span>' + node[i].cvName + subPageStr + '</span><\/div><ul tree-root-state="closed" id="tree-root' + node[i].id + '"><\/ul>';
+			html += '<img src="' + CCM_IMAGE_PATH + '/spacer.gif" width="16" height="16" class="handle ' + moveableClass + '" /><a href="javascript:toggleSub(' + node[i].id + ')"><img src="' + CCM_IMAGE_PATH + '/dashboard/plus.jpg" width="9" height="9" class="tree-plus" id="tree-collapse' + node[i].id + '" /><\/a><div rel="' + CCM_REL + '/index.php?cID=' + node[i].id + '" class="' + labelClass + '" tree-node-alias="' + cAlias + '" tree-node-canwrite="' + canWrite + '" tree-node-children="' + escape(node[i].numSubpages) + '" tree-node-title="' + escape(node[i].cvName) + '" id="tree-label' + node[i].id + '"><span>' + node[i].cvName + subPageStr + '</span><\/div><form onsubmit="return searchSitemapNode(' + node[i].id + ')" id="ccm-tree-search' + node[i].id + '" class="ccm-tree-search"><a href="javascript:void(0)" onclick="closeSitemapSearch(' + node[i].id + ')" class="ccm-tree-search-close"><img src="' + CCM_IMAGE_PATH + '/icons/close.png" /></a><input type="text" name="submit" name="q" /> <a href="javascript:void(0)" onclick="searchSitemapNode(' + node[i].id + ')"><img src="' + CCM_IMAGE_PATH + '/icons/magnifying.png" /></a></form><ul tree-root-state="closed" id="tree-root' + node[i].id + '"><\/ul>';
 		} else {
-			html += '<div tree-node-title="' + escape(node[i].cvName) + '" class="' + labelClass + '" tree-node-alias="' + cAlias + '" tree-node-canwrite="' + canWrite + '" id="tree-label' + node[i].id + '" rel="' + CCM_REL + '/index.php?cID=' + node[i].id + '"><img src="' + CCM_IMAGE_PATH + '/spacer.gif" width="16" height="16" class="handle ' + moveableClass + '" /><span>' + node[i].cvName + '</span><\/div>';
+			html += '<div tree-node-title="' + escape(node[i].cvName) + '" tree-node-children="' + escape(node[i].numSubpages) + '" class="' + labelClass + '" tree-node-alias="' + cAlias + '" tree-node-canwrite="' + canWrite + '" id="tree-label' + node[i].id + '" rel="' + CCM_REL + '/index.php?cID=' + node[i].id + '"><img src="' + CCM_IMAGE_PATH + '/spacer.gif" width="16" height="16" class="handle ' + moveableClass + '" /><span>' + node[i].cvName + '</span><\/div>';
 		}
 		html += '<\/li><div class="dropzone tree-dz' + nodeID + '" tree-parent="' + nodeID + '" id="tree-dz' + node[i].id + '"><\/div>';
 		
@@ -294,50 +311,50 @@ parseTree = function(node, nodeID, deactivateSubNodes) {
 	
 }
 
-selectLabel = function(e, node) {
-	switch(CCM_SITEMAP_MODE) {
-		case "move_copy_delete":
-			var destCID = node.attr('id').substring(10);
-			var origCID = CCM_CID;
-				
-			$.fn.dialog.open({
-				title: ccmi18n_sitemap.moveCopyPage,
-				href: CCM_TOOLS_PATH + '/dashboard/sitemap_drag_request.php?origCID=' + origCID + '&destCID=' + destCID,
-				width: 350,
-				modal: false,
-				height: 350,
-				onClose: function() {
-					//$("#tree").fadeIn(200);
-					if (tr_moveCopyMode == true) {
-						deactivateMoveCopy();
-					}
-				}
-
-			});
-			break;
-		case "select_page":
-
-			ccm_selectSitemapNode(node.attr('id').substring(10), unescape(node.attr('tree-node-title')));
-			jQuery.fn.dialog.closeTop();
-
-			break;
-		default:
-			node.addClass('tree-label-selected');
-			if (tr_activeNode != false) {
-				if (tr_activeNode.attr('id') != node.attr('id')) {
-					tr_activeNode.removeClass('tree-label-selected');
-				}
+selectMoveCopyTarget = function(destCID) {
+	var origCID = CCM_CID;
+		
+	$.fn.dialog.open({
+		title: ccmi18n_sitemap.moveCopyPage,
+		href: CCM_TOOLS_PATH + '/dashboard/sitemap_drag_request.php?origCID=' + origCID + '&destCID=' + destCID,
+		width: 350,
+		modal: false,
+		height: 350,
+		onClose: function() {
+			//$("#tree").fadeIn(200);
+			if (tr_moveCopyMode == true) {
+				deactivateMoveCopy();
 			}
-			params = {'cID': node.attr('id').substring(10), 'canWrite': node.attr('tree-node-canwrite'), 'cAlias': node.attr('tree-node-alias')};
-			showPageMenu(params, e);
-			tr_activeNode = node;
+		}
+
+	});
+}
+
+selectLabel = function(e, node) {
+	var cNumChildren = node.attr('tree-node-children');
+	if (CCM_SITEMAP_MODE == "move_copy_delete" && cNumChildren < 1) {
+		var destCID = node.attr('id').substring(10);
+		selectMoveCopyTarget(destCID);
+	} else if (CCM_SITEMAP_MODE == 'select_page') {
+		ccm_selectSitemapNode(node.attr('id').substring(10), unescape(node.attr('tree-node-title')));
+		jQuery.fn.dialog.closeTop();
+	} else {
+		node.addClass('tree-label-selected');
+		if (tr_activeNode != false) {
+			if (tr_activeNode.attr('id') != node.attr('id')) {
+				tr_activeNode.removeClass('tree-label-selected');
+			}
+		}
+		params = {'cID': node.attr('id').substring(10), 'canWrite': node.attr('tree-node-canwrite'), 'cNumChildren': node.attr('tree-node-children'), 'cAlias': node.attr('tree-node-alias')};
+		showPageMenu(params, e);
+		tr_activeNode = node;
 	}
 }
 
 activateLabels = function() {
-	$('div.tree-label').unbind();
-	$('div.tree-label').click(function(e) {
-		selectLabel(e, $(this))
+	$('div.tree-label span').unbind();
+	$('div.tree-label span').click(function(e) {
+		selectLabel(e, $(this).parent())
 	}); 
 }
 
@@ -388,6 +405,19 @@ moveCopyAliasNode = function(reloadPage) {
 	});	
 }
 
+searchSitemapNode = function(cID) {
+	var q = $('form#ccm-tree-search' + cID + ' input').val();
+	openSubSearch(cID, q);
+	return false;
+}
+
+closeSitemapSearch = function(cID) {
+	closeSub(cID);
+	var container = $("#tree-root" + cID);
+	$("#ccm-tree-search" + cID).hide();
+	container.removeClass('ccm-sitemap-search-results');
+}
+
 toggleSub = function(nodeID) {
 	ccm_hideMenus();
 	var container = $("#tree-root" + nodeID);
@@ -425,6 +455,24 @@ openSub = function(nodeID, onComplete) {
 		}, 200);
 	});	
 }
+
+openSubSearch = function(nodeID, query, onComplete) {
+	setLoading(nodeID);
+	var container = $("#tree-root" + nodeID);
+	container.addClass('ccm-sitemap-search-results');
+	cancelReorder();
+	$.getJSON(CCM_TOOLS_PATH + "/dashboard/sitemap_data.php?node=" + nodeID, {'keywords': query}, function(resp) {
+		parseTree(resp, nodeID, false);	
+		activateLabels();
+		setTimeout(function() {
+			removeLoading(nodeID);
+			if (onComplete != null) {
+				onComplete();
+			}			
+		}, 200);
+	});	
+}
+
 
 closeSub = function(nodeID) {
 	var container = $("#tree-root" + nodeID);
@@ -608,7 +656,7 @@ $(function() {
 		alert(ccmi18n_sitemap.loadError + request.responseText);
 	});
 	
-	$.getJSON(CCM_TOOLS_PATH + "/dashboard/sitemap_data.php", function(resp) {  
+	$.getJSON(CCM_TOOLS_PATH + "/dashboard/sitemap_data.php", {'mode' : CCM_SITEMAP_MODE}, function(resp) {  
 		parseTree(resp, 0, false);
 		activateLabels();
 		tr_doAnim = true;
